requirements.txt
Django==6.0.1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
psycopg[binary]==3.2.6
django-axes==6.5.0
django-encrypted-model-fields==0.6.5
django-cloudinary-storage==0.3.0
cloudinary==1.41.0
pyotp==2.9.0
qrcode==7.4.2
Pillow==11.1.0
reportlab==4.1.0
chardet==5.2.0
django-cors-headers==4.7.0
python-decouple==3.8

